# Optimize McKay decomposer