The main optimization is the replacement of `np.cross` with a direct calculation for 3d vectors. 

```
# on develop
%timeit myCircuit.decompose_mckay()
1.24 ms ± 54 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
# this PR
In [4]: %timeit myCircuit.decompose_mckay()
686 µs ± 79 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

@pablolh