@eendebakpt 2023-12-12
> I like your PR more for the nice code improvements than for the performance gain (we target feature and API over performance for now). The replacement of `np.cross` for pure performance reasons is debatable to me - you add more code and the performance gain will likely not be noticeable (this algo takes roughly linear time in the number of gates).


Fair argument. Let me know if you want me to revert the changes to `np.cross`, I do not have a strong opinion here.