@pablolh 2023-12-12
> Fair argument. Let me know if you want me to revert the changes to np.cross, I do not have a strong opinion here.

Let's revert the `np.cross` change then :)
I'll keep in mind that this `np.cross` seems to not be optimal performance-wise (for short vectors?).