@eendebakpt 2023-12-12
In general numpy has a very large overhead for short arrays (or scalars). For example:
```
> import math
>  import numpy
> 
> %timeit math.sin(.2)
> %timeit numpy.sin(.2)
102 ns ± 8.93 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
710 ns ± 37.5 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```
And `np.cross` in particular has a lot of overhead because it is much more general than a simple 3d vector cross.