@eendebakpt 2023-12-12
> > In general numpy has a very large overhead for short arrays (or scalars).
> 
> That is very good to know indeed!
> 
> In this particular case I am using numpy for the mathematical functions it provides, not for its compact arrays and fast C implementation of linear algebra. So probably indeed not the best use case for numpy, since this is just a 3d vector... Do you know an alternative library for this? We could implement cross product with native Python types ourselves but I also feel this is already implemented somewhere... with tests and assertions and corner cases and...
> 
> (I am also using numpy in a more standard way when I build the circuit matrix, which is kind of circuit simulation, therefore performance-heavy and appropriate for numpy.)

Unfortunately, there is no good lightweight (e.g. performant for small arrays) alternative to numpy. There is https://github.com/kwant-project/tinyarray, but the arrays are immutable. One can use cython or numba to increase performance, or use extensions such as rust.